+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 
 - Initial directories
 - CMakeLists.txt to store general data
+- Tensor.h for initial testing
 
 [unreleased]: https://github.com/IrishFix/UniversalNumerics/compare/v0.0.1...HEAD
 [0.0.1]: https://github.com/IrishFix/UniversalNumerics/releases/tag/v0.0.1